Seed for Products

+++ InventoryManagement.Application/Models/CompanyModel.cs
@@ -3,6 +3,6 @@
     public class CompanyModel : BaseModel
     {
         public string Name { get; set; }
-        public int Prefix { get; set; }
+        public long Prefix { get; set; }
     }
 }
