+++ InventoryManagement.Core/Entities/Company.cs
@@ -3,6 +3,6 @@
     public class Company : Entity
     {
         public string Name { get; set; }
-        public int Prefix { get; set; }
+        public long Prefix { get; set; }
     }
 }
